Re-add support for IAM instance profiles.
This was originally added in 9863cd6 (PR #2), but since then, upstream have
switched from using github.com/mitchellh/goamz to the official AWS
library (9dee348). When we rebased against upstream, our changes got
lost with this switch.

This therefore re-introduces them based on the similar support we added
to the S3 resource[1].

[1] https://github.com/alphagov/paas-s3-resource/pull/1